+++ NEWS.md
@@ -4,7 +4,8 @@ Patch to fix CRAN warning on Mac M1.
 
 ## Minor changes
 
-
+- Move away from `shiny:::processDeps` in favor of `htmltools::renderTags`. Does not
+impact end user. 
 
 # bs4Dash 2.0.1